Trekker ut arbeidsperiodefunksjoner til egen hook
